Extract ImageHandling at DSL level
