+++ shoes-core/lib/shoes/dsl.rb
@@ -60,6 +60,7 @@ require 'shoes/common/changeable'
 require 'shoes/common/clickable'
 require 'shoes/common/fill'
 require 'shoes/common/hover'
+require 'shoes/common/image_handling'
 require 'shoes/common/link_finder'
 require 'shoes/common/positioning'
 require 'shoes/common/remove'
