+++ shoes-core/lib/shoes/image.rb
@@ -3,6 +3,7 @@ class Shoes
   class Image < Common::UIElement
     include Common::Clickable
     include Common::Hover
+    include Common::ImageHandling
 
     BINARY_ENCODING = Encoding.find('binary')
 
@@ -39,27 +40,5 @@ class Shoes
       absolute_file_path(path_or_data)
     end
 
-    def absolute_file_path(path)
-      if Pathname(path).absolute?
-        search_for(File.basename(path), File.dirname(path))
-      else
-        search_for(path, *default_search_paths)
-      end
-    end
-
-    def search_for(path, *locations)
-      found = locations.map  { |dir| File.join(dir, path) }
-                       .find { |candidate| File.exist?(candidate) }
-
-      unless found
-        raise FileNotFoundError, "#{path} not found. Searched #{locations.join(',')}"
-      end
-
-      found
-    end
-
-    def default_search_paths
-      [Dir.pwd, Shoes.configuration.app_dir, File.join(Shoes::DIR, "static")]
-    end
   end
 end
